+++ pkg/armrpc/hostoptions/hostoptions.go
@@ -17,7 +17,7 @@ import (
 	v1 "github.com/project-radius/radius/pkg/armrpc/api/v1"
 	"github.com/project-radius/radius/pkg/azure/armauth"
 	aztoken "github.com/project-radius/radius/pkg/azure/tokencredentials"
-	"github.com/project-radius/radius/pkg/rp/kube"
+	"github.com/project-radius/radius/pkg/kubeutil"
 	"github.com/project-radius/radius/pkg/sdk"
 	"github.com/project-radius/radius/pkg/ucp/config"
 	sdk_cred "github.com/project-radius/radius/pkg/ucp/credentials"
@@ -136,7 +136,7 @@ func WithContext(ctx context.Context, cfg *ProviderConfig) context.Context {
 }
 
 func getKubernetes() (*rest.Config, error) {
-	cfg, err := kube.GetConfig()
+	cfg, err := kubeutil.NewClusterConfig()
 	if err != nil {
 		return nil, fmt.Errorf("failed to get kubernetes config: %w", err)
 	}
