+++ pkg/ucp/secret/provider/factory.go
@@ -9,7 +9,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/project-radius/radius/pkg/rp/kube"
+	"github.com/project-radius/radius/pkg/kubeutil"
 	"github.com/project-radius/radius/pkg/ucp/dataprovider"
 	"github.com/project-radius/radius/pkg/ucp/secret"
 	"github.com/project-radius/radius/pkg/ucp/secret/etcd"
@@ -37,14 +37,14 @@ func initETCDSecretClient(ctx context.Context, opts SecretProviderOptions) (secr
 	}
 	secretClient, ok := client.(*etcdstore.ETCDClient)
 	if !ok {
-		return nil, errors.New("No etcd Client detected")
+		return nil, errors.New("no etcd Client detected")
 	}
 	return &etcd.Client{ETCDClient: secretClient.Client()}, nil
 }
 
 func initKubernetesSecretClient(ctx context.Context, opt SecretProviderOptions) (secret.Client, error) {
 	s := scheme.Scheme
-	config, err := kube.GetConfig()
+	config, err := kubeutil.NewClusterConfig()
 	if err != nil {
 		return nil, err
 	}
